Split error.h into a proper header/source pair
I hesitated at first because the plan was to keep error.h a simple
header with a few macros, but that was before I learned that C89 does
not have variadic macros, so I had to use functions. For a while I
thought I could get away with simply making the functions static, but
it's an ugly solution and I don't like it.