Almost finish tokenization function
I got a lot of work done and it's mostly working. The only flaw I know
of right now is that query beginnings and ends are not checked properly
(i.e. you can begin/end a query with an operator). I'm hoping to fix
that by introducing more tokens and changing the flow of the function a
bit, but in case I destroy something I want to have a backup.