Add ArgLists, finish runQueries()
A LOT had to be added (fortunately not that much had to be changed) to
write the ini file parsing stage in a sane way, and it's still a bit
ugly. All additions have been thoroughly documented, but here's a brief
rundown:

- added ArgList as a counterpart to DataSet which stores actual values,
not section/key addresses of those values
- added IniToken type for easy representation of information fetched
from an INI file
- "finished" runQueries function (with NO TESTING, because it's
impossible to test without other components working, which I'll get done
in the near future)
- added a few helper functions and enums (getLine, IniLineType,
iniExtractFromLine etc.). Some of these have yet to be finalized and
properly connected and tested.

SIDE NOTE: Some of the newly introduced types have member names that
nest very badly (value->value->...), so once everything is up and
working I'll need to reformat them to make some passages more readable.