Complete INI file value extraction
The only program component missing right now is simply evaluation of
queries based on op_stack and the ArgLists. I've done basic testing on a
simple INI file and everything seems to be working well (valgrind shows
no memory leaks, and I left out a bunch of debug printing because I'll
be needing it for final testing which should come soon).

The side note from the previous commit holds and I will need to make
some serious name changes, because nested dereferencing is unreadable.

On a different note, I've implemented a distiction between INT and FLOAT
types, but I'm not sure if it was necessary, since using double for
every operation may be preferable and simpler.