Finish printQueries, most (all?) features work
- finished printQueries
- added a new stack type (ValStack) for storing ArgVals
- patched a few memory leaks (valgrind reports no problems)
- other minor changes

REMARK:
There is a few problems with the current evaluation approach. Each query
needs its own copy of each string that it processes, even if those
strings are the same values in the actual file. A good way to fix this
would be to aggregate values for all queries in a single shared
structure, and make each query store POINTERS to that shared structure
instead. I'm probably going to do that, along with other enhancements,
in future snapshots.