Clean rewrite of tokenizeQueryString
This snapshot changes the syntax rules for queries - now all section/key
pairs need to be enclosed in {curly braces}. The reason behind this
change was to allow sections and keys to use the '-' hyphen character
without introducing any ambiguity.

I took a different approach to the tokenization process, and to my utter
delight, the resulting code is much cleaner and much more extensible.

I even added implicit multiplication as a feature, so it is now legal to
write something like "{var1}{var2}" without an asterisk in-between (this
works for parentheses too).