Fix critical infixPostfix indexing mistake
For some reason I was using negative indexing in a small part of the
infix to postfix conversion function, and this issue was present for a
couple of days without me noticing it (or producing runtime errors!).

Also refactored the code for handling length-0 string multiplication
results (originally added a few snapshots back, but I found a more
concise and shorter way to do the same thing).